:sparkles: 메시지 읽음 처리 추가
