+++ src/components/layout/NavBar.tsx
@@ -3,10 +3,9 @@ import MessageDB from "@/server/prisma/message.db";
 import UserDB from "@/server/prisma/user.db";
 import AuthUtil from "@/server/utils/auth-util";
 import Bars3Icon from "@heroicons/react/24/solid/Bars3Icon";
-import BellIcon from "@heroicons/react/24/solid/BellIcon";
 import Link from "next/link";
 import ServerToastHelper from "../server-toast-helper";
-import classnames from "@/utils/classnames";
+import NavBarMessages from "./navbar-messages";
 
 export default async function NavBar() {
   const userSeq = await AuthUtil.getUserSeqThrows();
@@ -22,7 +21,6 @@ export default async function NavBar() {
   }
 
   const messages = await MessageDB.getMessagesByUser(userSeq);
-  const messageNotReadCount = messages.filter((it) => !it.isRead).length;
 
   return (
     <nav className="w-full navbar bg-base-100 fixed z-50">
@@ -73,47 +71,7 @@ export default async function NavBar() {
           </ul>
         </div>
 
-        <div className="dropdown dropdown-end dropdown-bottom">
-          <button tabIndex={1} className="btn btn-ghost btn-circle">
-            {messageNotReadCount ? (
-              <div className="indicator">
-                <BellIcon className="size-6" />
-                <span className="badge badge-xs badge-primary indicator-item">
-                  {messageNotReadCount}
-                </span>
-              </div>
-            ) : (
-              <BellIcon className="size-6" />
-            )}
-          </button>
-
-          <ul
-            tabIndex={1}
-            className="menu menu-sm dropdown-content mt-3 z-[1] p-2 shadow bg-base-100 rounded-box w-52"
-          >
-            {messages.map((message) => (
-              <li key={message.seq}>
-                <button
-                  className={classnames(
-                    "btn justify-start h-auto",
-                    message.isRead ? "" : "btn-ghost"
-                  )}
-                >
-                  <strong className="font-semibold text-sm">
-                    {message.title}
-                  </strong>
-                  {":"}
-                  <p className="font-normal text-xs">{message.content}</p>
-                </button>
-              </li>
-            ))}
-            <li>
-              <button className="btn btn-outline btn-neutral text-sm h-auto min-h-fit">
-                더보기
-              </button>
-            </li>
-          </ul>
-        </div>
+        <NavBarMessages messages={messages} />
       </div>
     </nav>
   );
